more refactoring the raycasting portion
