+++ src/plugins/lightrays.js
@@ -3,14 +3,12 @@ import PolyBool from 'polybooljs';
 
 import tilemapLayerToTileClumps from './clumpy';
 import tileToPolygon from './poly';
+import { getRaySegmentIntersection } from './rays';
 
 export default class LightraysPlugin extends Phaser.Plugins.BasePlugin {
   constructor(scene) {
     super('LightraysPlugin', scene);
-
-    console.log('-----------------------');
     this.scene = scene;
-    console.log(scene);
   }
 
   init(name) {
@@ -89,41 +87,6 @@ export default class LightraysPlugin extends Phaser.Plugins.BasePlugin {
     return polygonSegments;
   }
 
-  //kind of voodoo here.
-  getRaySegmentIntersection(ray, segment) {
-    let r_px = ray.a.x;
-    let r_py = ray.a.y;
-    let r_dx = ray.b.x - ray.a.x;
-    let r_dy = ray.b.y - ray.a.y;
-    let s_px = segment.a.x;
-    let s_py = segment.a.y;
-    let s_dx = segment.b.x - segment.a.x;
-    let s_dy = segment.b.y - segment.a.y;
-    let r_mag = Math.sqrt(r_dx * r_dx + r_dy * r_dy);
-    let s_mag = Math.sqrt(s_dx * s_dx + s_dy * s_dy);
-    if (r_dx / r_mag == s_dx / s_mag && r_dy / r_mag == s_dy / s_mag) {
-      return null;
-    }
-
-    let T2 =
-      (r_dx * (s_py - r_py) + r_dy * (r_px - s_px)) /
-      (s_dx * r_dy - s_dy * r_dx);
-    let T1 = (s_px + s_dx * T2 - r_px) / r_dx;
-
-    if (T1 < 0) {
-      return null;
-    }
-    if (T2 < 0 || T2 > 1) {
-      return null;
-    }
-
-    return {
-      x: r_px + r_dx * T1,
-      y: r_py + r_dy * T1,
-      param: T1
-    };
-  }
-
   //TODO: simplify this so there's less mapping
   createPolygonLayerFromTilemapLayer({ tilemapLayer, level }) {
     let { tilesets } = level;
@@ -141,17 +104,6 @@ export default class LightraysPlugin extends Phaser.Plugins.BasePlugin {
 
     let allPts = this.segmentsToPoints(this.occlusionSegments);
     this.occlusionPoints = this.pointsToUniquePoints(allPts);
-  }
-
-  bindLightToMouse() {
-    this._lightsource =
-      this._lightsource || new Phaser.Geom.Circle(200, 300, 4);
-
-    const game = this.scene.game;
-
-    //  only move when you click
-    const { worldX, worldY } = game.input.mousePointer;
-    this._lightsource.setPosition(worldX, worldY);
   }
 
   drawRay({ a, b }) {
@@ -181,57 +133,6 @@ export default class LightraysPlugin extends Phaser.Plugins.BasePlugin {
     return points.filter(toUniq);
   }
 
-  uniquePointsToIntersects(uniquePoints, { x, y }) {
-    let intersects = [];
-
-    let segments = this.occlusionSegments;
-    let tolerance = 0.00001;
-    var len = uniquePoints.length;
-
-    var uniqueAngles = [];
-    for (var j = 0; j < len; j++) {
-      var uniquePoint = uniquePoints[j];
-      var angle = Math.atan2(uniquePoint.y - y, uniquePoint.x - x);
-      uniquePoint.angle = angle;
-      uniqueAngles.push(angle - tolerance, angle, angle + tolerance);
-    }
-
-    var anglesLength = uniqueAngles.length;
-    for (var j = 0; j < anglesLength; j++) {
-      var angle = uniqueAngles[j];
-      // Calculate dx & dy from angle
-      var dx = Math.cos(angle);
-      var dy = Math.sin(angle);
-      // Ray from center of screen to mouse
-      var ray = {
-        a: { x: x, y: y },
-        b: { x: x + dx, y: y + dy }
-      };
-      // Find CLOSEST intersection
-      var closestIntersect = null;
-      for (var i = 0; i < segments.length; i++) {
-        var intersect = this.getRaySegmentIntersection(ray, segments[i]);
-        if (!intersect) continue;
-        if (!closestIntersect || intersect.param < closestIntersect.param) {
-          closestIntersect = intersect;
-        }
-      }
-      // Intersect angle
-      if (!closestIntersect) {
-        continue;
-      }
-      closestIntersect.angle = angle;
-      // Add to list of intersects
-      intersects.push(closestIntersect);
-    }
-
-    intersects = intersects.sort((a, b) => {
-      return a.angle - b.angle;
-    });
-
-    return intersects;
-  }
-
   sun = {
     worldX: 555,
     worldY: 49
@@ -254,10 +155,14 @@ export default class LightraysPlugin extends Phaser.Plugins.BasePlugin {
 
     this.lineGraphics.clear();
 
-    let beam = this.uniquePointsToIntersects(this.occlusionPoints, {
-      x: worldX,
-      y: worldY
-    });
+    let beam = uniquePointsToIntersects(
+      this.occlusionPoints,
+      this.occlusionSegments,
+      {
+        x: worldX,
+        y: worldY
+      }
+    );
 
     let beams = [beam];
     var fuzzyRadius = 8;
@@ -266,8 +171,9 @@ export default class LightraysPlugin extends Phaser.Plugins.BasePlugin {
     for (var angle = 0; angle < p2; angle += incrementByAngle) {
       var dx = Math.cos(angle) * fuzzyRadius;
       var dy = Math.sin(angle) * fuzzyRadius;
-      let sympatheticBeam = this.uniquePointsToIntersects(
+      let sympatheticBeam = uniquePointsToIntersects(
         this.occlusionPoints,
+        this.occlusionSegments,
         {
           x: worldX + dx,
           y: worldY + dy
